fire

+++ Assets/Scripts/BlockFactory.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class BlockFactory : MonoBehaviour
 {
-    public static void Factory(GameObject block, int x, int z)
+    public static void Factory(GameObject block, float x, float z)
     {
         Instantiate(block, new Vector3(x, 0, z), Quaternion.identity);
     }
